proxy/requirements.txt
solace-pubsubplus
requests
